option: reduce default number for TCP CT and NAT table max entries
Commit e824a86bba21 ("daemon: Allow configuration of CT max entries")
bumped the default value to 1000000 in order to ease upgrades from
Cilium 1.2. In the helm charts, the value was again set to 512KB via the
`ct-global-max-entries-tcp` option. However, if Cilium is not deployed
via helm charts (e.g. when running as a systemd service in the devel
VM) the large default number of entries is used.

Set the default value for `bpf-ct-global-tcp-max` to 512KB again and
instead advise users in the helm chart comments to set it to 1000000 in
case they're upgrading or changed the size manually using Helm.

Since the default value of `bpf-nat-global-max` for the NAT table size
is derived from the default for `bpf-ct-global-tcp-max`, this commit
will also decrease the the NAT table size to 512K.

Document possible consequences of upgrading Cilium installations with
larger TCP CT and NAT table sizes.

This saves about ~150MB of memory at runtime.

Updates #10056

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>