loader: Fix tunneling when device is set without NodePort
Many tests are failing when configuring a device while BPF-based NodePort
is disabled (cf. #11972). A bug in the loader causes it to call init.sh
with the mode set to direct instead of e.g., vxlan. The cilium_vxlan is
then not created (or even removed if it existed) which causes
connectivity disruptions in multiple scenarios (mostly multi-node).

This commit fixes the wrong assumption that multiple devices are only
set when NodePort is enabled. That assumption might have been correct
when that code was introduced (in 3a83623).

Fixes: #11972
Fixes: 3a83623 ("bpf: add support for local NodePort via tunnel")
Signed-off-by: Paul Chaignon <paul@cilium.io>