cilium, testing: sync encryption test options with non-encryption tests
Use the same options for the encryption+routing test that is used in the
test that does just routing. This gets the encryption+routing test passing
in my local cluster and in gke.

Then on vxlan test set the tunnel type so it reads the same as the other VXLAN
test.

Signed-off-by: John Fastabend <john.fastabend@gmail.com>